refacto(admin): move CollegueLogEntry to gsl_historique section
fix(admin): fix reverse URL for CollegueLogEntry changelist
